# packages/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Jomboy Lavka API
  version: 1.0.0
  description: Unified API for Jomboy Lavka client applications
servers:
  - url: https://api.jomboy-lavka.uz/api/v1
    description: Production
  - url: http://localhost:4010/api/v1
    description: Mock server
security:
  - bearerAuth: []
paths:
...
  /payments/initiate:
    post:
      tags: [Billing]
      summary: Initiate payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id, amount, provider]
              properties:
                order_id:
                  type: string
                  format: uuid
                amount:
                  type: integer
                provider:
                  type: string
                  enum: [payme, click]
      responses:
        '200':
          description: Payment initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment_id:
                    type: string
                  redirect_url:
                    type: string
                  provider:
                    type: string
  /refunds:
    post:
      tags: [Billing]
      summary: Process refund (Support role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id, amount, reason, ticket_id]
              properties:
                order_id:
                  type: string
                amount:
                  type: integer
                reason:
                  type: string
                ticket_id:
                  type: string
                type:
                  type: string
                  enum: [full, partial]
      responses:
        '200':
          description: Refund processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  refund_id:
                    type: string
                  status:
                    type: string
                    enum: [pending, completed, failed]
...
  /admin/warehouse/receipts:
    post:
      tags: [Warehouse]
      summary: Start goods receipt (Phase 2 demo)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                po_number:
                  type: string
      responses:
        '201':
          description: Receipt started
          content:
            application/json:
              schema:
                type: object
                properties:
                  receipt_id:
                    type: string
                  expected_items:
                    type: array
                    items:
                      type: object
                      properties:
                        sku_id:
                          type: string
                        name:
                          type: string
                        barcode:
                          type: string
                        expected_qty:
                          type: integer
...
  /admin/warehouse/writeoffs/{writeoff_id}/approve:
    post:
      tags: [Warehouse]
      summary: Director signature on writeoff
      parameters:
        - name: writeoff_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approved
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...